Renamed helpers to adapters
