+++ src/main/java/ch/zhaw/psit4/services/implementation/ConfigServiceImpl.java
@@ -14,8 +14,8 @@ import ch.zhaw.psit4.domain.exceptions.ZipFileCreationException;
 import ch.zhaw.psit4.domain.interfaces.DialPlanConfigurationInterface;
 import ch.zhaw.psit4.domain.interfaces.SipClientConfigurationInterface;
 import ch.zhaw.psit4.domain.sipclient.SipClientConfigurationChanSip;
-import ch.zhaw.psit4.services.implementation.helper.DialPlanConfigHelper;
-import ch.zhaw.psit4.services.implementation.helper.SipClientConfigHelper;
+import ch.zhaw.psit4.services.implementation.adapters.DialPlanConfigAdapter;
+import ch.zhaw.psit4.services.implementation.adapters.SipClientConfigAdapter;
 import ch.zhaw.psit4.services.interfaces.ConfigServiceInterface;
 import org.springframework.stereotype.Service;
 
@@ -35,14 +35,14 @@ import java.util.List;
  */
 @Service
 public class ConfigServiceImpl implements ConfigServiceInterface {
-    private final SipClientConfigHelper sipClientConfigHelper;
-    private final DialPlanConfigHelper dialPlanConfigHelper;
+    private final SipClientConfigAdapter sipClientConfigAdapter;
+    private final DialPlanConfigAdapter dialPlanConfigAdapter;
     private SipClientConfigurationInterface sipClientConfiguration;
     private DialPlanConfigurationInterface dialPlanConfiguration;
 
     public ConfigServiceImpl(SipClientRepository sipClientRepository, CompanyRepository companyRepository) {
-        this.dialPlanConfigHelper = new DialPlanConfigHelper(sipClientRepository, companyRepository);
-        this.sipClientConfigHelper = new SipClientConfigHelper(sipClientRepository);
+        this.dialPlanConfigAdapter = new DialPlanConfigAdapter(sipClientRepository, companyRepository);
+        this.sipClientConfigAdapter = new SipClientConfigAdapter(sipClientRepository);
         sipClientConfiguration = new SipClientConfigurationChanSip();
         dialPlanConfiguration = new DialPlanConfigurationChanSip();
     }
@@ -76,7 +76,7 @@ public class ConfigServiceImpl implements ConfigServiceInterface {
     public ByteArrayOutputStream getAsteriskConfiguration() {
         ConfigWriter configWriter = new ConfigWriter(sipClientConfiguration, dialPlanConfiguration);
 
-        List<SipClient> sipClientList = sipClientConfigHelper.getSipClientList();
+        List<SipClient> sipClientList = sipClientConfigAdapter.getSipClientList();
         List<DialPlanContext> contexts = getDialPlanContexts();
 
         String sipClientConf = configWriter.generateSipClientConfiguration(sipClientList);
@@ -88,9 +88,9 @@ public class ConfigServiceImpl implements ConfigServiceInterface {
     }
 
     private List<DialPlanContext> getDialPlanContexts() {
-        List<DialPlanContext> dialPlanContextList = dialPlanConfigHelper.getDialPlanContextList();
+        List<DialPlanContext> dialPlanContextList = dialPlanConfigAdapter.getDialPlanContextList();
 
-        List<Company> companyList = dialPlanConfigHelper.getCompanyDomainList();
+        List<Company> companyList = dialPlanConfigAdapter.getCompanyDomainList();
         //TODO probably it would be better if this call is in the domain, and a list of companies is passed to the domain
         List<DialPlanContext> defaultContexts = ContextGenerator.getDefaultContexts(companyList);
