+++ src/main/java/ch/zhaw/psit4/services/implementation/adapters/SipClientConfigAdapter.java
@@ -1,4 +1,4 @@
-package ch.zhaw.psit4.services.implementation.helper;
+package ch.zhaw.psit4.services.implementation.adapters;
 
 import ch.zhaw.psit4.data.jpa.repositories.SipClientRepository;
 import ch.zhaw.psit4.domain.beans.SipClient;
@@ -13,10 +13,10 @@ import static ch.zhaw.psit4.services.implementation.ConfigServiceImpl.sipClientE
  *
  * @author Jona Braun
  */
-public class SipClientConfigHelper {
+public class SipClientConfigAdapter {
     private final SipClientRepository sipClientRepository;
 
-    public SipClientConfigHelper(SipClientRepository sipClientRepository) {
+    public SipClientConfigAdapter(SipClientRepository sipClientRepository) {
         this.sipClientRepository = sipClientRepository;
     }
